check the item stock before create transaction

+++ delete.php
@@ -1,16 +1,20 @@
 <?php
-	require_once('includes/config.inc.php');
 
-	$md = $_GET['mid'];
-	
-	//connect Database
-	$link = mysqli_connect(DB_HOSTNAME, DB_USERNAME, DB_PASSWORD) or die("Could not connect to host");
-	mysqli_select_db($link, DB_DATABASE) or die("Could not find database");
+require_once('includes/config.inc.php');
+session_start();
+if ($_SESSION['uid'] == 1) {
+    $md = $_GET['mid'];
+    //connect Database
+    $link = mysqli_connect(DB_HOSTNAME, DB_USERNAME, DB_PASSWORD) or die("Could not connect to host");
+    mysqli_select_db($link, DB_DATABASE) or die("Could not find database");
 
-        $sql = "DELETE FROM `moviestore`.`movie` WHERE `movie`.`mid` = '".$md."'";
-	
-	$result = mysqli_query($link, $sql) or die("Data not found");
+    $sql = "DELETE FROM `moviestore`.`movie` WHERE `movie`.`mid` = '" . $md . "'";
 
-	header("location:adminpanel.php");
+    $result = mysqli_query($link, $sql) or die("Data not found");
+
+    header("location:adminpanel.php");
+}else{
+    header("location:index.php");
+}
 ?>
